Update View and BoardView to remove precompilation error
Removed the drawBoard() method from the interface View. Private methods cannot be declared in an interface.
By default it was being treated as public. Thus, it was not possible to assign the private modifier to
drawBoard() in the BoardView implementation of the View interface. The original JavaDoc content was saved and
added to JavaDoc comment for drawBoard in BoardView.