Changes to BoardModel and BoardView [Proposed changes]
Changes to BoardModel were focused on writing play(). play() was renamed
and now it seems better in another class. Consider moving paintComponent
to a controller. In other words the controller extends JPanel. Also, move
the startGame() to the controller. This seems to hold better benefits.
Changes to BoardView were focused on cleaning up comments.