Correct BoardModel and BoardView design errors
BoardModel and BoardView contained slight errors that, while allowing compilation, were
not inline with the project design.
BoardModel createGame was renamed play, and the while loop was removed. Additionally, it now
has an instance variable for UndoStructure.
BoardView was incorrectly extending JPanel and now extends JLabel.
Focus is now going to be on getting BoardView to 100% completion.